moved read write to new serialization sub-module
